refactor(note-access): remove redundant metadata config file
