+++ packages/note-access/src/metadata/constructor.js
@@ -1,4 +1,4 @@
-import metadataConfig, { START_OFFSET } from '../config/noteMetaData';
+import metadataConfig, { START_OFFSET } from '../config/metaData';
 import decodeMetaDataToObject from '../utils/decodeMetaDataToObject';
 import _addAccess from './_addAccess';
 import _getAccess from './_getAccess';
